+++ app/components/layout/menu.tsx
@@ -3,10 +3,10 @@ import { Ellipsis, LogOut } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { ScrollArea } from "@/components/ui/scroll-area";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
-import { getMenuList } from "@/lib/menu-list";
 import { cn } from "@/lib/utils";
 import { Link, useFetcher, useLocation } from "@remix-run/react";
 import { CollapseMenuButton } from "./collapse-menu-button";
+import { getMenuList } from "@/config.shared";
 
 interface MenuProps {
 	isOpen: boolean | undefined;
